Fix issues regarding thread safety of keras model in classification server
This commit fixes an error that occurred in the tensorflow
cpu version when trying to predict with a keras model in
a multithreaded context (namely in ClassificationServer).

The server now creates a session that is used over all
threads. Graph.finalize() is called to make it read-only.
Additionaly the ClassifierModel constructor needs to create the
prediction before the first thread is started.