Refactor logic for model creating, loading and training
This commit moves the logic for creating and loading a classifier model
from the model class itself to two separate classes. This allows classifiers
to be used easier in the classification server.

Furthermore models are now loaded via json files that contain all
relevant parameters. This has been done in order to avoid having
to hard code certain parameters or having to rely on too many
command line arguments. A similar mechanism is also used for
the training configuration.

Additional changes:
- Rename folder model to classifier
- Reorder scripts in classifier folder
- Replace hard coded model config and port in classification server with
  command line arguments