Add support for features other than word indices
This commit adds the option to specify a "feature_extractor" when
loading or creating a model, as opposed to having word indices
hard-coded.

A consequence of this change is that the model factory no longer knows
the length of the input, which is why it needs to specified explicitly as
a parameter.