Increase documents used when testing memory usage
Use 100 documents when testing memory usage instead of one.
This should prevent the process from terminating before the profiler is
attached.

As a result of this change predict.py now accepts arbitrarily many
document paths. This approach is currently limited by the character
limit for command line arguments, which is around 32000 on Windows 10.